fix(naming): schema to attribute and name to key
Code in the form field (singular) used the schema property to store an attribute
of a schema. This is confusing.

The name property of an attribute was changed to key. This better fits existing
standards. It also makes sense as a keyword that represents a property is often
refered to as key.

Also updated version